+++ programs/MushroomClassifier.php
@@ -47,8 +47,11 @@ $dataset = new Labeled($samples, $labels);
 
 $estimator = new Pipeline(new MultiLayerPerceptron([
     new Dense(30, new SELU()),
-    new AlphaDropout(20, new SELU(), 0.1),
+    new AlphaDropout(0.1),
+    new Dense(20, new SELU()),
+    new AlphaDropout(0.1),
     new Dense(10, new SELU()),
+    new AlphaDropout(0.05),
 ], 100, new Adam(0.001), 1e-4, new CrossEntropy(), 1e-3, new MCC(), 0.1, 3, 100), [
     new OneHotEncoder(),
     new SparseRandomProjector(30),