+++ src/NeuralNet/Layers/AlphaDropout.php
@@ -5,8 +5,6 @@ namespace Rubix\ML\NeuralNet\Layers;
 use MathPHP\LinearAlgebra\Matrix;
 use MathPHP\LinearAlgebra\MatrixFactory;
 use Rubix\ML\NeuralNet\Optimizers\Optimizer;
-use Rubix\ML\NeuralNet\ActivationFunctions\SELU;
-use Rubix\ML\NeuralNet\ActivationFunctions\ActivationFunction;
 use InvalidArgumentException;
 use RuntimeException;
 
@@ -23,14 +21,12 @@ use RuntimeException;
  * @package     Rubix/ML
  * @author      Andrew DalPino
  */
-class AlphaDropout extends Dense
+class AlphaDropout implements Hidden, Nonparametric
 {
     const ALPHA = 1.6732632423543772848170429916717;
     const SCALE = 1.0507009873554804934193349852946;
 
     const ALPHA_P = -self::ALPHA * self::SCALE;
-
-    const PHI = 1000000;
 
     /**
      * The ratio of neurons that are dropped during each training pass.
@@ -54,6 +50,13 @@ class AlphaDropout extends Dense
     protected $b;
 
     /**
+     * The width of the layer.
+     *
+     * @var int
+     */
+    protected $width;
+
+    /**
      * The memoized dropout mask.
      *
      * @var \MathPHP\LinearAlgebra\Matrix|null
@@ -61,29 +64,42 @@ class AlphaDropout extends Dense
     protected $mask;
 
     /**
-     * @param  int  $neurons
-     * @param  \Rubix\ML\NeuralNet\ActivationFunctions\ActivationFunction  $activationFunction
      * @param  float  $ratio
      * @throws \InvalidArgumentException
      * @return void
      */
-    public function __construct(int $neurons, ActivationFunction $activationFunction = null,
-                                float $ratio = 0.1)
+    public function __construct(float $ratio = 0.1)
     {
         if ($ratio <= 0.0 or $ratio >= 1.0) {
             throw new InvalidArgumentException('Dropout ratio must be between 0'
                 . ' and 1.0.');
         }
 
-        if (is_null($activationFunction)) {
-            $activationFunction = new SELU(self::ALPHA, self::SCALE);
-        }
-
         $this->ratio = $ratio;
         $this->a = ((1 - $ratio) * (1 + $ratio * self::ALPHA_P ** 2)) ** -0.5;
         $this->b = -$this->a * self::ALPHA_P * $ratio;
+        $this->width = 0;
+    }
 
-        parent::__construct($neurons, $activationFunction);
+    /**
+     * @return int
+     */
+    public function width() : int
+    {
+        return $this->width;
+    }
+
+    /**
+     * Initialize the layer.
+     *
+     * @param  int  $fanIn
+     * @return int
+     */
+    public function init(int $fanIn) : int
+    {
+        $this->width = $fanIn;
+
+        return $fanIn;
     }
 
     /**
@@ -95,12 +111,8 @@ class AlphaDropout extends Dense
      */
     public function forward(Matrix $input) : Matrix
     {
-        $this->input = $input;
-
-        $this->z = $this->weights->w()->multiply($input);
-
-        $m = $this->z->getM() - 1;
-        $n = $this->z->getN();
+        $m = $input->getM();
+        $n = $input->getN();
 
         $mask = MatrixFactory::zero($m, $n)->map(function ($value) {
             return (rand(0, self::PHI) / self::PHI) > $this->ratio ? 1.0 : 0.0;
@@ -110,52 +122,49 @@ class AlphaDropout extends Dense
             return $value === 0.0 ? self::ALPHA_P : 0.0;
         });
 
-        $biases = MatrixFactory::one(1, $n);
-
-        $temp = $this->z->rowExclude($m);
-
-        $this->computed = $this->activationFunction->compute($temp)
-            ->hadamardProduct($mask)
+        $activations = $input->hadamardProduct($mask)
             ->add($saturation)
             ->map(function ($activation) {
                 return $this->a * $activation + $this->b;
-            })
-            ->augmentBelow($biases);
+            });
 
-        $this->mask = $mask->augmentBelow($biases);
+        $this->mask = $mask;
 
-        return $this->computed;
+        return $activations;
+    }
+
+    /**
+     * Compute the inferential activations of each neuron in the layer.
+     *
+     * @param  \MathPHP\LinearAlgebra\Matrix  $input
+     * @return \MathPHP\LinearAlgebra\Matrix
+     */
+    public function infer(Matrix $input) : Matrix
+    {
+        return $input;
     }
 
     /**
      * Calculate the errors and gradients of the layer and update the parameters.
      *
-     * @param  \MathPHP\LinearAlgebra\Matrix  $prevWeights
-     * @param  \MathPHP\LinearAlgebra\Matrix  $prevErrors
+     * @param  callable  $prevErrors
      * @param  \Rubix\ML\NeuralNet\Optimizers\Optimizer  $optimizer
      * @throws \RuntimeException
-     * @return array
+     * @return callable
      */
-    public function back(Matrix $prevWeights, Matrix $prevErrors, Optimizer $optimizer) : array
+    public function back(callable $prevErrors, Optimizer $optimizer) : callable
     {
-        if (is_null($this->input) or is_null($this->z) or is_null($this->computed) or is_null($this->mask)) {
+        if (is_null($this->mask)) {
             throw new RuntimeException('Must perform forward pass before'
                 . ' backpropagating.');
         }
 
-        $errors = $this->activationFunction
-            ->differentiate($this->z, $this->computed)
-            ->hadamardProduct($prevWeights->transpose()->multiply($prevErrors))
-            ->hadamardProduct($this->mask);
+        $errors = call_user_func($prevErrors)->hadamardProduct($this->mask);
 
-        $gradient = $errors->multiply($this->input->transpose());
+        unset($this->mask);
 
-        $step = $optimizer->step($this->weights, $gradient);
-
-        $this->weights->update($step);
-
-        unset($this->input, $this->z, $this->computed, $this->mask);
-
-        return [$this->weights->w(), $errors];
+        return function () use ($errors) {
+            return $errors;
+        };
     }
 }