+++ src/NeuralNet/Layers/Nonparametric.php
@@ -1,0 +1,24 @@
+<?php
+
+namespace Rubix\ML\NeuralNet\Layers;
+
+use MathPHP\LinearAlgebra\Matrix;
+
+interface Nonparametric extends Layer
+{
+    /**
+     * Compute the training activations of each neuron in the layer.
+     *
+     * @param  \MathPHP\LinearAlgebra\Matrix  $input
+     * @return \MathPHP\LinearAlgebra\Matrix
+     */
+    public function forward(Matrix $input) : Matrix;
+
+    /**
+     * Compute the inferential activations of each neuron in the layer.
+     *
+     * @param  \MathPHP\LinearAlgebra\Matrix  $input
+     * @return \MathPHP\LinearAlgebra\Matrix
+     */
+    public function infer(Matrix $input) : Matrix;
+}