+++ tests/NeuralNet/Layers/DropoutTest.php
@@ -3,9 +3,9 @@
 namespace Rubix\Tests\NeuralNet\Layers;
 
 use Rubix\ML\NeuralNet\Layers\Layer;
-use Rubix\ML\NeuralNet\Layers\Dropout;
 use Rubix\ML\NeuralNet\Layers\Hidden;
-use Rubix\ML\NeuralNet\Layers\Parametric;
+use Rubix\ML\NeuralNet\Layers\Dropout;
+use Rubix\ML\NeuralNet\Layers\Nonparametric;
 use Rubix\ML\NeuralNet\ActivationFunctions\ELU;
 use PHPUnit\Framework\TestCase;
 
@@ -15,7 +15,7 @@ class DropoutTest extends TestCase
 
     public function setUp()
     {
-        $this->layer = new Dropout(10, new ELU(), 0.5);
+        $this->layer = new Dropout(0.5);
     }
 
     public function test_build_layer()
@@ -23,11 +23,13 @@ class DropoutTest extends TestCase
         $this->assertInstanceOf(Dropout::class, $this->layer);
         $this->assertInstanceOf(Layer::class, $this->layer);
         $this->assertInstanceOf(Hidden::class, $this->layer);
-        $this->assertInstanceOf(Parametric::class, $this->layer);
+        $this->assertInstanceOf(Nonparametric::class, $this->layer);
     }
 
     public function test_width()
     {
-        $this->assertEquals(11, $this->layer->width());
+        $this->layer->init(10);
+
+        $this->assertEquals(10, $this->layer->width());
     }
 }